Name change mwcore -> core
