+++ src/core/CrossCorrelation.h
