+++ src/core/FilterCache.cpp
@@ -8,7 +8,7 @@
  * \breif
  */
 
-#include "mwcore/FilterCache.h"
+#include "core/FilterCache.h"
 #include "mwutils/Printer.h"
 
 #include "constants.h"
