+++ src/core/GaussQuadrature.h
