+++ src/core/MWFilter.h
