+++ src/core/ObjectCache.h
