+++ tests/factory_functions.h
@@ -4,8 +4,8 @@
 #include "mwtrees/BoundingBox.h"
 #include "mwtrees/NodeIndex.h"
 #include "mwtrees/MultiResolutionAnalysis.h"
-#include "mwcore/InterpolatingBasis.h"
-#include "mwcore/LegendreBasis.h"
+#include "core/InterpolatingBasis.h"
+#include "core/LegendreBasis.h"
 #include "mwtrees/FunctionTree.h"
 #include "functions/GaussFunc.h"
 #include "mwutils/Printer.h"
